+++ examples/react-example/CHANGELOG.md
@@ -1,5 +1,17 @@
 # react-example
 
+## 2.8.5
+
+### Patch Changes
+
+- @stackframe/react@2.8.5
+
+## 2.8.4
+
+### Patch Changes
+
+- @stackframe/react@2.8.4
+
 ## 2.8.3
 
 ### Patch Changes